+++ mlf_core/create/templates/mlflow/mlflow_xgboost/{{ cookiecutter.project_slug }}/{{ cookiecutter.project_slug_no_hyphen }}/{{ cookiecutter.project_slug_no_hyphen }}.py
@@ -3,6 +3,7 @@ import xgboost as xgb
 import mlflow
 import mlflow.xgboost
 import time
+import GPUtil
 
 from rich import traceback
 
@@ -14,10 +15,15 @@ from data_loading.data_loader import load_train_test_data
 @click.option('--epochs', type=int, default=5, help='Number of epochs to train')
 @click.option('--general-seed', type=int, default=0, help='General Python, Python random and Numpy seed.')
 @click.option('--xgboost-seed', type=int, default=0, help='XGBoost specific random seed.')
-@click.option('--cuda', type=click.Choice(['True', 'False']), help='Enable or disable CUDA support.')
+@click.option('--cuda', type=click.Choice(['True', 'False']), default=True, help='Enable or disable CUDA support.')
 @click.option('--single-precision-histogram', default=True, help='Enable or disable single precision histogram calculation.')
 def start_training(epochs, general_seed, xgboost_seed, cuda, single_precision_histogram):
-    use_cuda = True if cuda == 'True' else False
+    avail_gpus = GPUtil.getGPUs()
+    use_cuda = True if cuda == 'True' and len(avail_gpus) > 0 else False
+    if use_cuda:
+        click.echo(click.style(f'Using {len(avail_gpus)} GPUs!', fg='blue'))
+    else:
+        click.echo(click.style(f'No GPUs detected. Running on the CPU', fg='blue'))
 
     with mlflow.start_run():
         # Fetch and prepare data
@@ -27,24 +33,23 @@ def start_training(epochs, general_seed, xgboost_seed, cuda, single_precision_hi
         mlflow.xgboost.autolog()
 
         # Set XGBoost parameters
-        param = {
-            'objective': 'multi:softmax',
-            'num_class': 8,
-        }
-        param['single_precision_histogram'] = True if single_precision_histogram == 'True' else False
-        param['subsample'] = 0.5
-        param['colsample_bytree'] = 0.5
-        param['colsample_bylevel'] = 0.5
+        param = {'objective': 'multi:softmax',
+                 'num_class': 8,
+                 'single_precision_histogram': True if single_precision_histogram == 'True' else False,
+                 'subsample': 0.5,
+                 'colsample_bytree': 0.5,
+                 'colsample_bylevel': 0.5}
 
         # Set random seeds
         set_general_random_seeds(general_seed)
         set_xgboost_random_seeds(xgboost_seed, param)
 
         # Set CPU or GPU as training device
-        if not use_cuda:
-            param['tree_method'] = 'hist'
-        else:
+        if use_cuda:
             param['tree_method'] = 'gpu_hist'
+        else:
+            param['tree_method'] = 'hist'
+        print(param['tree_method'])
 
         # Train on the chosen device
         results = {}